#872 Fixed scipy 1.5 not always keeping the dtype of the original array during getitem (especially if the result is an empty array).

+++ src/secml/array/c_sparse.py
@@ -423,8 +423,9 @@ class CSparse(_CArrayInterface):
         # Check index for all other cases
         idx = self._check_index(idx)
 
-        # Ready for scipy
-        return self.__class__(self._data.__getitem__(idx))
+        # Ready for scipy. We cast the result of scipy's getitem as the
+        # original dtype is not kept sometimes (especially for empty arrays)
+        return self.__class__(self._data.__getitem__(idx), dtype=self.dtype)
 
     def __setitem__(self, idx, value):
         """Redefinition of the get (brackets) operator."""
